sppm to context_module
